Conflicts:
	Rakefile
	configure
	kernel/bootstrap/load_order.txt
	kernel/common/array.rb
	kernel/common/load_order.txt
	kernel/delta/load_order.txt
	kernel/platform/library.rb
	rakelib/blueprint.rb
	rakelib/vm.rake
	spec/tags/18/ruby/core/kernel/eval_tags.txt
	spec/tags/18/ruby/core/kernel/public_methods_tags.txt
	spec/tags/18/ruby/library/fiber/transfer_tags.txt
	spec/tags/18/ruby/library/iconv/iconv_tags.txt
	spec/tags/20/ruby/library/socket/basicsocket/close_read_tags.txt
	spec/tags/20/ruby/library/socket/basicsocket/close_write_tags.txt
	spec/tags/20/ruby/library/uri/eql_tags.txt
	spec/tags/20/ruby/library/uri/equality_tags.txt
	vm/builtin/class.cpp
	vm/builtin/nativefunction.cpp
	vm/builtin/nativemethod.cpp
	vm/builtin/regexp.cpp
	vm/builtin/string.cpp
	vm/builtin/system.cpp
	vm/builtin/taskprobe.cpp
	vm/capi/string.cpp
	vm/codegen/field_extract.rb
	vm/environment.cpp
	vm/external_libs/winpthreads/tests/ChangeLog
	vm/gc/managed.cpp
	vm/gc/managed.hpp
	vm/gc/slab.hpp
	vm/instruments/profiler.cpp
	vm/instruments/profiler.hpp
	vm/llvm/jit_util.cpp
	vm/llvm/jit_visit.hpp
	vm/native_thread.cpp
	vm/objectmemory.cpp
	vm/ontology.cpp
	vm/primitives.cpp
	vm/shared_state.cpp
	vm/shared_state.hpp
	vm/test/test_profiler.hpp
	vm/vm.cpp
	vm/vm.hpp
	vm/vmmethod.cpp

+++ lib/ext/openssl/ossl_pkey.c
@@ -1,5 +1,5 @@
 /*
- * $Id: ossl_pkey.c 12496 2007-06-08 15:02:04Z technorama $
+ * $Id$
  * 'OpenSSL for Ruby' project
  * Copyright (C) 2001-2002  Michal Rokos <m.rokos@sh.cvut.cz>
  * All rights reserved.
@@ -177,7 +177,7 @@ ossl_pkey_sign(VALUE self, VALUE digest, VALUE data)
     str = rb_str_new(0, EVP_PKEY_size(pkey)+16);
     if (!EVP_SignFinal(&ctx, RSTRING_PTR(str), &buf_len, pkey))
 	ossl_raise(ePKeyError, NULL);
-    assert(buf_len <= RSTRING_LEN(str));
+    assert((long)buf_len <= RSTRING_LEN(str));
     rb_str_set_len(str, buf_len);
 
     return str;
